+++ src/core/settings.h
@@ -10,7 +10,7 @@
 #include <unordered_map>
 #include <vector>
 #include "common/common_types.h"
-#include "core/hle/service/cam/cam.h"
+#include "core/hle/service/cam/cam_params.h"
 
 namespace Settings {
